backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
uvicorn==0.24.0
django-redis==5.4.0
django-filter==23.5
django-extensions==3.2.3
python-dateutil==2.8.2
requests==2.31.0
cryptography==41.0.7
